feat!: migrate public API from f32 to f64 for TBP numerical precision (#11)
* feat!: migrate public API from f32 to f64 for TBP numerical precision

BREAKING CHANGE: All public API types now use f64 for TBP Python NumPy compatibility

- ObjectRotation: pitch/yaw/roll fields changed from f32 to f64
- CameraIntrinsics: focal_length/principal_point changed from [f32;2] to [f64;2]
- RenderOutput.depth: changed from Vec<f32> to Vec<f64>
- RenderConfig::intrinsics(): returns f64 values
- get_depth(): returns Option<f64>
- to_depth_image(): returns Vec<Vec<f64>>
- unproject(): accepts f64 depth, returns [f64;3]

Bevy internal types (Transform, Vec3, Quat) remain f32, with conversion
at the boundary via to_quat() which converts f64 to f32 for Bevy.

Serialization format for fixtures maintains f32 for backward compatibility,
with conversion to f64 when loading.

Closes #10



* docs: update CLAUDE.md with f64 types and release flow documentation

- Document release-please workflow for automated versioning
- Update API documentation to reflect f64 types
- Add guidance on conventional commits for breaking changes



---------

+++ examples/test_render.rs
@@ -48,10 +48,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     println!("Unique colors in image: {}", unique_colors.len());
 
     // Check depth variation (should not be uniform if real depth readback works)
-    let unique_depths: std::collections::HashSet<u32> =
+    let unique_depths: std::collections::HashSet<u64> =
         output.depth.iter().map(|d| d.to_bits()).collect();
-    let depth_min = output.depth.iter().cloned().reduce(f32::min);
-    let depth_max = output.depth.iter().cloned().reduce(f32::max);
+    let depth_min = output.depth.iter().cloned().reduce(f64::min);
+    let depth_max = output.depth.iter().cloned().reduce(f64::max);
     println!(
         "Unique depth values: {} (range: {:?} to {:?})",
         unique_depths.len(),
